requirements.txt
streamlit 
langchain-community
google-generativeai
